Add feign. Rename package.
