--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/config/PasswordEncoderConfig.java
@@ -1,15 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordEncoderConfig {
-
-    @Bean
-    public PasswordEncoder getPasswordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
-}
