--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/config/RestTempleConfig.java
@@ -1,16 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.config;
-
-import org.springframework.cloud.client.loadbalancer.LoadBalanced;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTempleConfig {
-
-    @LoadBalanced
-    @Bean
-    public RestTemplate getRestTemplate(){
-        return new RestTemplate();
-    }
-}
