--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/converter/Converter.java
@@ -1,6 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.converter;
-
-public interface Converter<R, T> {
-
-    R convert(T object);
-}
