--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/dto/UserDto.java
@@ -1,20 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserDto {
-
-    private String userId;
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-    private String encryptedPassword;
-}
