--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/filter/AuthenticationFilter.java
@@ -1,58 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.filter;
-
-import by.dulik.eureka.users.by.dulik.eureka.users.dto.UserDto;
-import by.dulik.eureka.users.by.dulik.eureka.users.model.LoginUserRequestDto;
-import by.dulik.eureka.users.by.dulik.eureka.users.service.UsersService;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.Date;
-
-@RequiredArgsConstructor(onConstructor = @__(@Autowired))
-public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
-
-    private final UsersService usersService;
-    private final Environment environment;
-
-    @SneakyThrows
-    @Override
-    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
-        LoginUserRequestDto cred = new ObjectMapper().readValue(request.getInputStream(), LoginUserRequestDto.class);
-        return getAuthenticationManager()
-                .authenticate(new UsernamePasswordAuthenticationToken(cred.getEmail(), cred.getPassword(), new ArrayList<>()));
-    }
-
-    @Override
-    protected void successfulAuthentication(HttpServletRequest request,
-                                            HttpServletResponse response,
-                                            FilterChain chain,
-                                            Authentication authResult) {
-
-        String email = ((User) authResult.getPrincipal()).getUsername();
-        UserDto userDto = usersService.getUserDetailsByEmail(email);
-
-        String token = Jwts.builder()
-                .setSubject(userDto.getUserId())
-                .setExpiration(new Date(System.currentTimeMillis() + Long.parseLong(environment.getProperty("token.expiration.time"))))
-                .signWith(SignatureAlgorithm.HS512, environment.getProperty("token.secret"))
-                .compact();
-
-        response.addHeader("token", token);
-        response.addHeader("userId", userDto.getUserId());
-    }
-
-}
