--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/mapper/util/EncodePassword.java
@@ -1,14 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.mapper.util;
-
-import org.mapstruct.Qualifier;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Qualifier
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.CLASS)
-public @interface EncodePassword {
-}
