--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/model/response/CreateUserResponseDto.java
@@ -1,18 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.model.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateUserResponseDto {
-
-    private String userId;
-    private String firstName;
-    private String lastName;
-    private String email;
-}
