--- users/src/main/java/by/dulik/eureka/users/by/dulik/eureka/users/model/response/MessageResponseDto.java
@@ -1,18 +1,0 @@
-package by.dulik.eureka.users.by.dulik.eureka.users.model.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class MessageResponseDto {
-
-    private String messageId;
-    private String userId;
-    private String title;
-    private String text;
-}
