+++ users/src/main/java/by/dulik/eureka/users/config/RestTempleConfig.java
@@ -1,0 +1,23 @@
+package by.dulik.eureka.users.config;
+
+import feign.Logger;
+import feign.Logger.Level;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTempleConfig {
+
+    @LoadBalanced
+    @Bean
+    public RestTemplate getRestTemplate() {
+        return new RestTemplate();
+    }
+
+    @Bean
+    public Logger.Level feignLogger(){
+        return Level.FULL;
+    }
+}
