+++ users/src/main/java/by/dulik/eureka/users/config/SecurityConfig.java
@@ -1,0 +1,51 @@
+package by.dulik.eureka.users.config;
+
+import by.dulik.eureka.users.filter.AuthenticationFilter;
+import by.dulik.eureka.users.service.UsersService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Value("${gateway.ip}")
+    public String pathApiGateway;
+
+    @Value("${login.url.path}")
+    public String loginUrlPath;
+
+    private final UsersService usersService;
+    private final Environment environment;
+    private final PasswordEncoder passwordEncoder;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+        http.authorizeRequests().antMatchers("/users/**").permitAll()
+                .and()
+                .addFilter(getAuthenticationFilter());
+        http.headers().frameOptions().disable();
+    }
+
+    private AuthenticationFilter getAuthenticationFilter() throws Exception {
+        AuthenticationFilter authenticationFilter = new AuthenticationFilter(usersService, environment);
+        authenticationFilter.setAuthenticationManager(authenticationManager());
+        authenticationFilter.setFilterProcessesUrl(environment.getProperty("login.url.path"));
+        return authenticationFilter;
+    }
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(usersService).passwordEncoder(passwordEncoder);
+    }
+}
