+++ users/src/main/java/by/dulik/eureka/users/converter/Converter.java
@@ -1,0 +1,6 @@
+package by.dulik.eureka.users.converter;
+
+public interface Converter<R, T> {
+
+    R convert(T object);
+}
