+++ users/src/main/java/by/dulik/eureka/users/exception/CustomUserApiException.java
@@ -1,0 +1,22 @@
+package by.dulik.eureka.users.exception;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import org.springframework.http.HttpStatus;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Builder
+@AllArgsConstructor
+@Data
+public class CustomUserApiException {
+
+    private HttpStatus status;
+    private String message;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy hh:mm:ss")
+    private LocalDateTime timestamp;
+    private List<String> errors;
+}
