+++ users/src/main/java/by/dulik/eureka/users/mapper/UserMapperTest.java
@@ -1,0 +1,52 @@
+package by.dulik.eureka.users.mapper;
+
+//import by.dulik.eureka.data.entity.User;
+//import by.dulik.eureka.users.dto.UserDto;
+//import org.junit.jupiter.api.Test;
+//import org.junit.jupiter.api.extension.ExtendWith;
+//import org.mockito.junit.jupiter.MockitoExtension;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.boot.test.context.SpringBootTest;
+//import org.springframework.test.context.ActiveProfiles;
+//import org.springframework.test.context.junit.jupiter.SpringExtension;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertNotNull;
+//
+//@ExtendWith({SpringExtension.class, MockitoExtension.class})
+//@ActiveProfiles("test")
+//@SpringBootTest
+class UserMapperTest {
+
+//    @Autowired
+//    private UserMapper userMapper;
+//
+//    @Test
+//    void userToUserDtoTest() {
+//        User user = User.builder()
+//                .userId("1")
+//                .firstName("fn")
+//                .lastName("ln")
+//                .email("e")
+//                .encryptedPassword("passEnc")
+//                .build();
+//        UserDto userDto = userMapper.userToUserDto(user);
+//        assertNotNull(userDto);
+//        assertEquals(user.getFirstName(), userDto.getFirstName());
+//    }
+//
+//    @Test
+//    void userDtoToUserTest(){
+//        UserDto userDto = UserDto.builder()
+//                .userId("1")
+//                .firstName("fn")
+//                .lastName("ln")
+//                .email("e")
+//                .encryptedPassword("passEnc")
+//                .build();
+//        User user = userMapper.userDtoToUser(userDto);
+//        assertNotNull(user);
+//        assertEquals(userDto.getFirstName(), user.getFirstName());
+//    }
+
+}
