+++ users/src/main/java/by/dulik/eureka/users/model/LoginUserRequestDto.java
@@ -1,0 +1,16 @@
+package by.dulik.eureka.users.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginUserRequestDto {
+
+    private String email;
+    private String password;
+}
