+++ users/src/main/resources/application.properties
@@ -19,3 +19,6 @@ spring.rabbitmq.host=localhost
 spring.rabbitmq.port=5672
 spring.rabbitmq.username=guest
 spring.rabbitmq.password=guest
+
+#Logging
+logging.level.by.dulik.eureka.users.service.MessageServiceClient=DEBUG
